requirements.txt
#aniso8601==8.0.0
#click==7.1.2
#Flask==1.1.2
#Flask-RESTful==0.3.8
#Flask-SQLAlchemy==2.4.3
#itsdangerous==1.1.0
#Jinja2==2.11.2
#MarkupSafe==1.1.1
#pytz==2020.1
#six==1.15.0
#SQLAlchemy==1.3.18
#Werkzeug==1.0.1
#requests==2.22.0
#pytest==6.2.3
#gunicorn==20.0.4
djangorestframework==3.12.4
Markdown==3.3.4
django-filter==2.4.0
requests==2.22.0
gunicorn==20.0.4